Update script due to review comments
